requirements.txt
dataset==1.6.0
datasets==2.10.1
gym==0.17.3
ignite==1.1.0
jsonlines==3.1.0
keras==2.11.0
lief==0.10.1
lightgbm==3.1.1
matplotlib==3.3.4
numpy==1.20.1
omegaconf==2.3.0
pandas==1.2.4
path.py==12.5.0
pefile==2022.5.30
pexpect==4.8.0
Pillow==9.4.0
psutil==5.8.0
requests==2.25.1
scikit_learn==1.2.2
scipy==1.6.2
tensorflow==2.11.0
torch==1.13.1
torch_geometric==2.2.0
torchvision==0.14.1
tqdm==4.59.0
visdom==0.2.4
